chain-lab/requirements.txt
openai
langchain
langchain-community
python-dotenv
